chore: rename lib/deskflow/protocol_types => lib/deskflow/ProtocolTypes
