+++ src/lib/client/HelloBack.cpp
@@ -7,8 +7,8 @@
 #include "HelloBack.h"
 
 #include "base/Log.h"
+#include "deskflow/ProtocolTypes.h"
 #include "deskflow/ProtocolUtil.h"
-#include "deskflow/protocol_types.h"
 
 #include <map>
 #include <set>
