+++ src/lib/client/ServerProxy.cpp
@@ -17,10 +17,10 @@
 #include "deskflow/ClipboardChunk.h"
 #include "deskflow/FileChunk.h"
 #include "deskflow/OptionTypes.h"
+#include "deskflow/ProtocolTypes.h"
 #include "deskflow/ProtocolUtil.h"
 #include "deskflow/StreamChunker.h"
 #include "deskflow/XDeskflow.h"
-#include "deskflow/protocol_types.h"
 #include "io/IStream.h"
 
 #include <algorithm>
