+++ src/lib/deskflow/ClientApp.cpp
@@ -16,9 +16,9 @@
 #include "common/constants.h"
 #include "deskflow/ArgParser.h"
 #include "deskflow/ClientArgs.h"
+#include "deskflow/ProtocolTypes.h"
 #include "deskflow/Screen.h"
 #include "deskflow/XScreen.h"
-#include "deskflow/protocol_types.h"
 #include "net/NetworkAddress.h"
 #include "net/SocketMultiplexer.h"
 #include "net/TCPSocketFactory.h"
