+++ src/lib/deskflow/FileChunk.cpp
@@ -9,8 +9,8 @@
 #include "base/Log.h"
 #include "base/Stopwatch.h"
 #include "base/String.h"
+#include "deskflow/ProtocolTypes.h"
 #include "deskflow/ProtocolUtil.h"
-#include "deskflow/protocol_types.h"
 #include "io/IStream.h"
 
 static const uint16_t kIntervalThreshold = 1;
