+++ src/lib/server/ClientProxyUnknown.cpp
@@ -11,9 +11,9 @@
 #include "base/Log.h"
 #include "base/TMethodEventJob.h"
 #include "deskflow/AppUtil.h"
+#include "deskflow/ProtocolTypes.h"
 #include "deskflow/ProtocolUtil.h"
 #include "deskflow/XDeskflow.h"
-#include "deskflow/protocol_types.h"
 #include "io/IStream.h"
 #include "io/XIO.h"
 #include "server/ClientProxy1_0.h"
