+++ src/lib/server/Server.cpp
@@ -19,10 +19,10 @@
 #include "deskflow/IPlatformScreen.h"
 #include "deskflow/OptionTypes.h"
 #include "deskflow/PacketStreamFilter.h"
+#include "deskflow/ProtocolTypes.h"
 #include "deskflow/Screen.h"
 #include "deskflow/StreamChunker.h"
 #include "deskflow/XDeskflow.h"
-#include "deskflow/protocol_types.h"
 #include "mt/Thread.h"
 #include "net/TCPSocket.h"
 #include "server/ClientListener.h"
